requirements.txt
praw==7.7.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2  # For Google Gemini (free tier)
groq==0.4.2  # For Groq (free tier)
ollama==0.1.7  # For local LLMs
transformers==4.36.0  # For Hugging Face models
torch==2.1.0  # For local models